chore: sample data moved to test folder
